Small fixes

+++ create_db.py
@@ -3,5 +3,6 @@ import os
 from app import app, db
 
 with app.app_context():
-    os.remove("instance/project.db")
+    if os.path.exists("instance/project.db"):
+        os.remove("instance/project.db")
     db.create_all()
